Cleanup slurm package
